improve project structure for home Table + Grid components
