+++ plugins/techdocs/src/home/components/Grids/DocsCardGrid.test.tsx
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import React from 'react';
-import { render } from '@testing-library/react';
-import { wrapInTestApp } from '@backstage/test-utils';
+
 import { configApiRef } from '@backstage/core-plugin-api';
-import { DocsTable } from './DocsTable';
-import { rootDocsRouteRef } from '../../routes';
-import { entityRouteRef } from '@backstage/plugin-catalog-react';
+import { wrapInTestApp } from '@backstage/test-utils';
+import { render } from '@testing-library/react';
+import React from 'react';
+import { rootDocsRouteRef } from '../../../routes';
+import { DocsCardGrid } from './DocsCardGrid';
 
 // Hacky way to mock a specific boolean config value.
 const getOptionalBooleanMock = jest.fn().mockReturnValue(false);
@@ -40,15 +40,15 @@ jest.mock('@backstage/core-plugin-api', () => ({
   },
 }));
 
-describe('DocsTable test', () => {
+describe('Entity Docs Card Grid', () => {
   beforeEach(() => {
     jest.resetAllMocks();
   });
 
-  it('should render documents passed', async () => {
-    const { findByText } = render(
+  it('should render all entities passed ot it', async () => {
+    const { findByText, findAllByRole } = render(
       wrapInTestApp(
-        <DocsTable
+        <DocsCardGrid
           entities={[
             {
               apiVersion: 'version',
@@ -57,19 +57,8 @@ describe('DocsTable test', () => {
                 name: 'testName',
               },
               spec: {
-                owner: 'user:owned',
+                owner: 'techdocs@example.com',
               },
-              relations: [
-                {
-                  target: {
-                    kind: 'user',
-                    namespace: 'default',
-                    name: 'owned',
-                  },
-                  targetRef: 'user:default/owned',
-                  type: 'ownedBy',
-                },
-              ],
             },
             {
               apiVersion: 'version',
@@ -80,37 +69,23 @@ describe('DocsTable test', () => {
               spec: {
                 owner: 'not-owned@example.com',
               },
-              relations: [
-                {
-                  target: {
-                    kind: 'user',
-                    namespace: 'default',
-                    name: 'not-owned',
-                  },
-                  targetRef: 'user:default/not-owned',
-                  type: 'ownedBy',
-                },
-              ],
             },
           ]}
         />,
         {
           mountedRoutes: {
             '/docs/:namespace/:kind/:name/*': rootDocsRouteRef,
-            '/catalog/:namespace/:kind/:name': entityRouteRef,
           },
         },
       ),
     );
-
-    const link1 = await findByText('testName');
-    const link2 = await findByText('testName2');
-    expect(link1).toBeInTheDocument();
-    expect(link1.getAttribute('href')).toContain(
+    expect(await findByText('testName')).toBeInTheDocument();
+    expect(await findByText('testName2')).toBeInTheDocument();
+    const [button1, button2] = await findAllByRole('button');
+    expect(button1.getAttribute('href')).toContain(
       '/docs/default/testkind/testname',
     );
-    expect(link2).toBeInTheDocument();
-    expect(link2.getAttribute('href')).toContain(
+    expect(button2.getAttribute('href')).toContain(
       '/docs/default/testkind2/testname2',
     );
   });
@@ -118,9 +93,9 @@ describe('DocsTable test', () => {
   it('should fall back to case-sensitive links when configured', async () => {
     getOptionalBooleanMock.mockReturnValue(true);
 
-    const { findByText } = render(
+    const { findByRole } = render(
       wrapInTestApp(
-        <DocsTable
+        <DocsCardGrid
           entities={[
             {
               apiVersion: 'version',
@@ -130,32 +105,20 @@ describe('DocsTable test', () => {
                 namespace: 'SomeNamespace',
               },
               spec: {
-                owner: 'user:owned',
+                owner: 'techdocs@example.com',
               },
-              relations: [
-                {
-                  target: {
-                    kind: 'user',
-                    namespace: 'default',
-                    name: 'owned',
-                  },
-                  targetRef: 'user:default/owned',
-                  type: 'ownedBy',
-                },
-              ],
             },
           ]}
         />,
         {
           mountedRoutes: {
             '/techdocs/:namespace/:kind/:name/*': rootDocsRouteRef,
-            '/catalog/:namespace/:kind/:name': entityRouteRef,
           },
         },
       ),
     );
 
-    const button = await findByText('testName');
+    const button = await findByRole('button');
     expect(getOptionalBooleanMock).toHaveBeenCalledWith(
       'techdocs.legacyUseCaseSensitiveTripletPaths',
     );
@@ -164,16 +127,31 @@ describe('DocsTable test', () => {
     );
   });
 
-  it('should render empty state if no owned documents exist', async () => {
+  it('should render entity title if available', async () => {
     const { findByText } = render(
-      wrapInTestApp(<DocsTable entities={[]} />, {
-        mountedRoutes: {
-          '/docs/:namespace/:kind/:name/*': rootDocsRouteRef,
-          '/catalog/:namespace/:kind/:name': entityRouteRef,
+      wrapInTestApp(
+        <DocsCardGrid
+          entities={[
+            {
+              apiVersion: 'version',
+              kind: 'TestKind',
+              metadata: {
+                name: 'testName',
+                title: 'TestTitle',
+              },
+              spec: {
+                owner: 'techdocs@example.com',
+              },
+            },
+          ]}
+        />,
+        {
+          mountedRoutes: {
+            '/docs/:namespace/:kind/:name/*': rootDocsRouteRef,
+          },
         },
-      }),
+      ),
     );
-
-    expect(await findByText('No documents to show')).toBeInTheDocument();
+    expect(await findByText('TestTitle')).toBeInTheDocument();
   });
 });
