+++ plugins/techdocs/src/home/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Backstage Authors
+ * Copyright 2022 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,4 @@
  * limitations under the License.
  */
 
-import { Entity, EntityName } from '@backstage/catalog-model';
-
-export type DocsTableRow = {
-  entity: Entity;
-  resolved: {
-    docsUrl: string;
-    ownedByRelationsTitle: string;
-    ownedByRelations: EntityName[];
-  };
-};
+export * from './components';
